Introducing Controllers
Controllers are added to run the business logic. This will reduce the lines of code on the routes which gets the Http Request and pass it to Controller which executes the actual logic.
