+++ app.js
@@ -7,6 +7,7 @@ const app = express();
 
 app.use('/api/places',placesRoutes);
 
+// Express Js will execute the  below custom function, when it senses an error by other middleware functions like app.use('/api/places',placesRoutes);
 app.use( (error, req, res, next) => {
     if(res.headerSent){
         return next(error);
